updated:       - URLParser       - LinkParser       - GitHubLinkParser       - StackOverflowLinkParser
   added:
       - Response
       - GitHubResponse
       - StackOverflowResponse
