+++ link-parser/src/main/java/ru/tinkoff/edu/java/parser/URLParser.java
@@ -1,12 +1,15 @@
-import ru.tinkoff.edu.java.parsers.LinkParser;
+package ru.tinkoff.edu.java.parser;
+
+import ru.tinkoff.edu.java.parser.parsers.LinkParser;
+import ru.tinkoff.edu.java.parser.responses.Response;
 
 
 public class URLParser {
 
-    public String parseLink(String link) {
+    public Response parse(String link) {
         @SuppressWarnings("unchecked") Class<LinkParser>[] d = (Class<LinkParser>[]) LinkParser.class.getPermittedSubclasses();
         for (Class<LinkParser> linkParserClass : d) {
-            String result = null;
+            Response result = null;
             try {
                 result = linkParserClass.getDeclaredConstructor().newInstance().parseLink(link);
             } catch (ReflectiveOperationException e) {
