+++ link-parser/src/main/java/ru/tinkoff/edu/java/parser/parsers/LinkParser.java
@@ -1,0 +1,7 @@
+package ru.tinkoff.edu.java.parser.parsers;
+
+import ru.tinkoff.edu.java.parser.responses.Response;
+
+public sealed interface LinkParser permits GitHubLinkParser, StackOverflowLinkParser {
+    Response parseLink(String link);
+}
