+++ link-parser/src/main/java/ru/tinkoff/edu/java/parser/responses/GitHubResponse.java
@@ -1,0 +1,4 @@
+package ru.tinkoff.edu.java.parser.responses;
+
+public record GitHubResponse(String userName, String repository) implements Response{
+}
