+++ link-parser/src/main/java/ru/tinkoff/edu/java/parser/responses/Response.java
@@ -1,0 +1,4 @@
+package ru.tinkoff.edu.java.parser.responses;
+
+public sealed interface Response permits GitHubResponse, StackOverflowResponse{
+}
