+++ link-parser/src/main/java/ru/tinkoff/edu/java/parser/responses/StackOverflowResponse.java
@@ -1,0 +1,4 @@
+package ru.tinkoff.edu.java.parser.responses;
+
+public record StackOverflowResponse(Integer id) implements Response{
+}
