--- link-parser/src/main/java/ru/tinkoff/edu/java/parsers/LinkParser.java
@@ -1,5 +1,0 @@
-package ru.tinkoff.edu.java.parsers;
-
-public sealed interface LinkParser permits GitHubLinkParser, StackOverflowLinkParser {
-    String parseLink(String link);
-}
